services/file-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
firebase-admin==6.2.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.11.0
python-multipart==0.0.6
python-magic==0.4.27
pydantic-settings==2.0.3
python-dotenv==1.0.0 